+++ libraries/test_session_current/src/main/java/androidx/media3/session/MockPlayer.java
@@ -633,6 +633,9 @@ public class MockPlayer implements Player {
       return;
     }
     boolean wasPlaying = isPlaying();
+    if (playbackState != STATE_IDLE) {
+      this.playerError = null;
+    }
     this.playbackState = playbackState;
     boolean isPlaying = isPlaying();
     for (Listener listener : listeners) {